+++ modules/AWSPowerShell/Cmdlets/Glue/Basic/Get-GLUEUnfilteredPartitionsMetadata-Cmdlet.cs
@@ -43,11 +43,21 @@ namespace Amazon.PowerShell.Cmdlets.GLUE
         #region Parameter AuditContext_AdditionalAuditContext
         /// <summary>
         /// <para>
-        /// The service has not provided documentation for this parameter; please refer to the service's API reference documentation for the latest available information.
+        /// <para>The context for the audit..</para>
         /// </para>
         /// </summary>
         [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
         public System.String AuditContext_AdditionalAuditContext { get; set; }
+        #endregion
+        
+        #region Parameter AuditContext_AllColumnsRequested
+        /// <summary>
+        /// <para>
+        /// <para>All columns request for audit.</para>
+        /// </para>
+        /// </summary>
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        public System.Boolean? AuditContext_AllColumnsRequested { get; set; }
         #endregion
         
         #region Parameter CatalogId
@@ -92,6 +102,17 @@ namespace Amazon.PowerShell.Cmdlets.GLUE
         /// </summary>
         [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
         public System.String Expression { get; set; }
+        #endregion
+        
+        #region Parameter AuditContext_RequestedColumn
+        /// <summary>
+        /// <para>
+        /// <para>The requested columns for audit.</para>
+        /// </para>
+        /// </summary>
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        [Alias("AuditContext_RequestedColumns")]
+        public System.String[] AuditContext_RequestedColumn { get; set; }
         #endregion
         
         #region Parameter Segment
@@ -220,6 +241,11 @@ namespace Amazon.PowerShell.Cmdlets.GLUE
             }
             #pragma warning restore CS0618, CS0612 //A class member was marked with the Obsolete attribute
             context.AuditContext_AdditionalAuditContext = this.AuditContext_AdditionalAuditContext;
+            context.AuditContext_AllColumnsRequested = this.AuditContext_AllColumnsRequested;
+            if (this.AuditContext_RequestedColumn != null)
+            {
+                context.AuditContext_RequestedColumn = new List<System.String>(this.AuditContext_RequestedColumn);
+            }
             context.CatalogId = this.CatalogId;
             #if MODULAR
             if (this.CatalogId == null && ParameterWasBound(nameof(this.CatalogId)))
@@ -287,6 +313,26 @@ namespace Amazon.PowerShell.Cmdlets.GLUE
             if (requestAuditContext_auditContext_AdditionalAuditContext != null)
             {
                 request.AuditContext.AdditionalAuditContext = requestAuditContext_auditContext_AdditionalAuditContext;
+                requestAuditContextIsNull = false;
+            }
+            System.Boolean? requestAuditContext_auditContext_AllColumnsRequested = null;
+            if (cmdletContext.AuditContext_AllColumnsRequested != null)
+            {
+                requestAuditContext_auditContext_AllColumnsRequested = cmdletContext.AuditContext_AllColumnsRequested.Value;
+            }
+            if (requestAuditContext_auditContext_AllColumnsRequested != null)
+            {
+                request.AuditContext.AllColumnsRequested = requestAuditContext_auditContext_AllColumnsRequested.Value;
+                requestAuditContextIsNull = false;
+            }
+            List<System.String> requestAuditContext_auditContext_RequestedColumn = null;
+            if (cmdletContext.AuditContext_RequestedColumn != null)
+            {
+                requestAuditContext_auditContext_RequestedColumn = cmdletContext.AuditContext_RequestedColumn;
+            }
+            if (requestAuditContext_auditContext_RequestedColumn != null)
+            {
+                request.AuditContext.RequestedColumns = requestAuditContext_auditContext_RequestedColumn;
                 requestAuditContextIsNull = false;
             }
              // determine if request.AuditContext should be set to null
@@ -408,6 +454,8 @@ namespace Amazon.PowerShell.Cmdlets.GLUE
         internal partial class CmdletContext : ExecutorContext
         {
             public System.String AuditContext_AdditionalAuditContext { get; set; }
+            public System.Boolean? AuditContext_AllColumnsRequested { get; set; }
+            public List<System.String> AuditContext_RequestedColumn { get; set; }
             public System.String CatalogId { get; set; }
             public System.String DatabaseName { get; set; }
             public System.String Expression { get; set; }